+++ custom_components/stopwatch_plus/__init__.py
@@ -28,7 +28,7 @@ from .services import async_setup_services
 from .source import SourceBinding
 from .stopwatch import Stopwatch
 
-PLATFORMS: list[Platform] = [Platform.BUTTON, Platform.SENSOR]
+PLATFORMS: list[Platform] = [Platform.BUTTON, Platform.EVENT, Platform.SENSOR]
 
 CONFIG_SCHEMA = cv.config_entry_only_config_schema(DOMAIN)